+++ libs/util-matrix/src/index.ts
@@ -19,6 +19,7 @@ export * from './lib/password-uia';
 export * from './lib/attachment-crypto';
 export * from './lib/key-file-crypto';
 export * from './lib/authed-media';
+export * from './lib/room-avatar';
 export * from './lib/room-create';
 export * from './lib/room-state';
 export * from './lib/crypto-wasm-loader';